refactor(web): decompose AuthPage.tsx (694 → <600 LOC) under max-lines budget
Split monolithic AuthPage.tsx into a composition shell + focused
subcomponents/hook/schemas under apps/web/src/core/auth/:

- AuthPage.tsx (150) — shell + mode-switch composition
- LoginForm.tsx (133)
- RegisterForm.tsx (152)
- ForgotPasswordPanel.tsx (85) — JSX panel
- useForgotPassword.ts (87) — panel state + 6-сек auto-close timer
- GoogleSignInButton.tsx (43)
- authFormPrimitives.tsx (99) — FieldError, PasswordStrengthBar,
  PasswordVisibilityToggle
- authSchemas.ts (38) — zod login/register schemas + types

No public-behavior changes. Public export AuthPage remains exported
from the same path. All new files <600 raw LOC and pass max-lines
(Hard Rule #18, sergeant-web-ui).

Bookkeeping:
- docs/tech-debt/frontend.md §4 — drop AuthPage.tsx from the leaker
  lookup table, bump 'Last validated' freshness marker.
- apps/web/eslint.i18n-allowlist.json — replace stale AuthPage.tsx
  entry with the new files that now hold the inline cyrillic literals
  (carry-over only; i18n burndown unchanged).
