+++ apps/web/eslint.i18n-allowlist.json
@@ -11,7 +11,12 @@
   "apps/web/src/core/app/MigrationPrompt.tsx",
   "apps/web/src/core/app/WelcomeScreen.tsx",
   "apps/web/src/core/auth/AuthPage.tsx",
+  "apps/web/src/core/auth/ForgotPasswordPanel.tsx",
+  "apps/web/src/core/auth/GoogleSignInButton.tsx",
+  "apps/web/src/core/auth/LoginForm.tsx",
+  "apps/web/src/core/auth/RegisterForm.tsx",
   "apps/web/src/core/auth/ResetPasswordPage.tsx",
+  "apps/web/src/core/auth/authFormPrimitives.tsx",
   "apps/web/src/core/components/CapabilityDetailModal.tsx",
   "apps/web/src/core/components/ChatInput.tsx",
   "apps/web/src/core/components/ChatMessage.tsx",
