+++ apps/web/src/core/auth/AuthPage.tsx
@@ -1,451 +1,31 @@
-import { useEffect, useRef, useState, type SyntheticEvent } from "react";
-import { z } from "zod";
-import { cn } from "@shared/lib/ui/cn";
+import { useState } from "react";
 import { Button } from "@shared/components/ui/Button";
 import { Card } from "@shared/components/ui/Card";
-import { Input } from "@shared/components/ui/Input";
 import { useCelebration } from "@shared/components/ui/CelebrationModal";
-import { useToast } from "@shared/hooks/useToast";
-import { useApiForm } from "@shared/forms/useApiForm";
-import { messages } from "@shared/i18n/uk";
-import { estimatePasswordStrength } from "@shared/lib/auth/passwordStrength";
 import { BrandLogo } from "../app/BrandLogo";
 import { useAuth } from "./AuthContext";
-
-// Зод-схеми тримаємо поряд з AuthPage, бо вони вузько-локальні (не
-// використовуються більше ніде). Окремий пакет `@sergeant/auth-schemas`
-// був би оверкіл-ом для двох форм. Меседжі — з `messages.validation.*`
-// (`apps/web/src/shared/i18n/uk.ts`), див. AGENTS.md (Hard Rule #15) і
-// `docs/i18n/readiness.md`.
-const loginSchema = z.object({
-  email: z
-    .string()
-    .min(1, messages.validation.emailRequired)
-    .email(messages.validation.emailInvalid),
-  // На login-у ми не нав'язуємо мінімальну довжину пароля — користувач
-  // міг створити акаунт у епоху 6-символьного мінімуму, а потім стандарт
-  // підняли. Перевірка відбувається на сервері; форма просто гарантує,
-  // що поле не порожнє.
-  password: z.string().min(1, messages.validation.passwordRequired),
-});
-type LoginValues = z.infer<typeof loginSchema>;
-
-const registerSchema = z.object({
-  email: z
-    .string()
-    .min(1, messages.validation.emailRequired)
-    .email(messages.validation.emailInvalid),
-  password: z
-    .string()
-    .min(10, messages.validation.passwordMin10)
-    // Better Auth-у достатньо просто довжини, але натякаємо
-    // користувачеві, що 10+ символів — нижня межа надійності.
-    .max(128, messages.validation.passwordMax128),
-  // Імʼя — опціональне; якщо не введене, fallback на `email.split("@")[0]`
-  // нижче в `onSubmit`. Залишаємо пустий рядок як валідне значення, щоб
-  // RHF не показав помилку «обовʼязкове поле» — це необовʼязкове.
-  name: z.string().max(80, messages.validation.nameMax80).optional(),
-});
-type RegisterValues = z.infer<typeof registerSchema>;
-
-function PasswordStrengthBar({ password }: { password: string }) {
-  if (!password) return null;
-  // PR-15 / §C8 — entropy-aware ladder. Замінює naive довжина-only оцінку,
-  // що однаково вважала надійним і `aaaaaaaaaa`, і `Aa1!Aa1!Aa`. Лейбли —
-  // bare-string (rule scope: тільки JSX-літерали), окремий i18n-namespace
-  // не виправдано для трьох коротких токенів.
-  const { level } = estimatePasswordStrength(password);
-  const widths = ["w-1/3", "w-2/3", "w-full"];
-  const colors = ["bg-error", "bg-amber-400", "bg-brand-500"];
-  const labels = ["Слабкий", "Середній", "Надійний"];
-  const labelColors = [
-    "text-error",
-    "text-amber-500",
-    "text-brand-strong dark:text-brand",
-  ];
-
-  return (
-    <div className="mt-1.5 space-y-1">
-      <div className="h-1 rounded-full bg-line overflow-hidden">
-        <div
-          className={cn(
-            "h-full rounded-full transition-all duration-300",
-            widths[level],
-            colors[level],
-          )}
-        />
-      </div>
-      <p className={cn("text-meta font-medium", labelColors[level])}>
-        {labels[level]}
-      </p>
-    </div>
-  );
-}
-
-interface PasswordVisibilityToggleProps {
-  visible: boolean;
-  onToggle: () => void;
-}
-
-function PasswordVisibilityToggle({
-  visible,
-  onToggle,
-}: PasswordVisibilityToggleProps) {
-  return (
-    <button
-      type="button"
-      onClick={onToggle}
-      aria-label={visible ? "Приховати пароль" : "Показати пароль"}
-      className="absolute right-3 top-1/2 -translate-y-1/2 text-muted hover:text-text transition-colors focus:outline-none focus-visible:ring-2 focus-visible:ring-brand-500/45 rounded"
-    >
-      {visible ? (
-        <svg
-          width="18"
-          height="18"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        >
-          <path d="M17.94 17.94A10.07 10.07 0 0112 20c-7 0-11-8-11-8a18.45 18.45 0 015.06-5.94" />
-          <path d="M9.9 4.24A9.12 9.12 0 0112 4c7 0 11 8 11 8a18.5 18.5 0 01-2.16 3.19" />
-          <line x1="1" y1="1" x2="23" y2="23" />
-        </svg>
-      ) : (
-        <svg
-          width="18"
-          height="18"
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        >
-          <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" />
-          <circle cx="12" cy="12" r="3" />
-        </svg>
-      )}
-    </button>
-  );
-}
-
-interface FieldErrorProps {
-  message: string | undefined;
-}
-
-function FieldError({ message }: FieldErrorProps) {
-  if (!message) return null;
-  return (
-    <p className="mt-1.5 text-meta text-error" role="alert">
-      {message}
-    </p>
-  );
-}
-
-interface LoginFormProps {
-  onForgotPassword: (currentEmail: string) => void;
-  showForgot: boolean;
-}
-
-function LoginForm({ onForgotPassword, showForgot }: LoginFormProps) {
-  const { login, authError } = useAuth();
-  const toast = useToast();
-  const [showPassword, setShowPassword] = useState(false);
-
-  // useApiForm зводить isSubmitting / валідацію / dirty-state в один
-  // hook. Серверні top-level-помилки сюди НЕ протікають — Better Auth
-  // повертає їх через `authError` з `useAuth()`, тож ми просто
-  // перекидаємо `Error("")` в `onSubmit`, щоб придушити `onSuccess`.
-  const { register, submit, formState, isSubmitting } = useApiForm<
-    LoginValues,
-    boolean
-  >({
-    schema: loginSchema,
-    defaultValues: { email: "", password: "" },
-    onSubmit: async (values) => {
-      const ok = await login(values.email, values.password);
-      if (!ok) {
-        // Кидаємо мовчазний error — він блокує `onSuccess`, але не
-        // показується (рендер `serverError` свідомо не приводимо).
-        // Реальний текст помилки відображається через `authError`
-        // нижче, бо він утримує локалізоване повідомлення Better Auth.
-        throw new Error("");
-      }
-      return ok;
-    },
-    onSuccess: () => {
-      toast.success("Вхід виконано");
-    },
-  });
-
-  // Стежимо за поточним email у полі — потрібен `<button "Забули пароль">`,
-  // щоб попередньо заповнити email у форму скидання пароля.
-  const emailValue = formState.defaultValues?.email ?? "";
-
-  return (
-    <form onSubmit={submit} noValidate className="space-y-4">
-      <div>
-        <label
-          htmlFor="auth-email"
-          className="block text-style-caption text-muted mb-1.5"
-        >
-          Email
-        </label>
-        <Input
-          id="auth-email"
-          type="email"
-          placeholder="email@example.com"
-          autoComplete="email"
-          error={!!formState.errors.email}
-          aria-invalid={!!formState.errors.email}
-          disabled={isSubmitting}
-          {...register("email")}
-        />
-        <FieldError message={formState.errors.email?.message} />
-      </div>
-
-      <div>
-        <div className="flex items-center justify-between mb-1.5">
-          <label
-            htmlFor="auth-password"
-            className="block text-style-caption text-muted"
-          >
-            Пароль
-          </label>
-          <button
-            type="button"
-            onClick={() => onForgotPassword(emailValue)}
-            className="text-xs text-brand-strong dark:text-brand-400 hover:underline focus:outline-none focus-visible:ring-2 focus-visible:ring-brand-500/45 rounded"
-          >
-            Забули пароль?
-          </button>
-        </div>
-        <div className="relative">
-          <Input
-            id="auth-password"
-            type={showPassword ? "text" : "password"}
-            placeholder="Пароль"
-            autoComplete="current-password"
-            className="pr-10"
-            error={!!formState.errors.password}
-            aria-invalid={!!formState.errors.password}
-            disabled={isSubmitting}
-            {...register("password")}
-          />
-          <PasswordVisibilityToggle
-            visible={showPassword}
-            onToggle={() => setShowPassword((v) => !v)}
-          />
-        </div>
-        <FieldError message={formState.errors.password?.message} />
-      </div>
-
-      {/* `authError` тримає локалізоване повідомлення з Better Auth
-          (translateAuthError). Не робимо `serverError` з useApiForm,
-          щоб не дублювати джерело істини — auth context сам володіє
-          серверною помилкою (login + Google + reset). */}
-      {authError && !showForgot && (
-        <div
-          role="alert"
-          className="text-xs text-error bg-error/10 border border-error/20 rounded-xl px-4 py-2.5"
-        >
-          {authError}
-        </div>
-      )}
-
-      <Button
-        type="submit"
-        variant="primary"
-        size="lg"
-        loading={isSubmitting}
-        className="w-full"
-      >
-        {isSubmitting ? messages.loadingActions.signingIn : "Увійти"}
-      </Button>
-    </form>
-  );
-}
-
-interface RegisterFormProps {
-  onAlreadyRegistered: () => void;
-}
-
-function RegisterForm({ onAlreadyRegistered }: RegisterFormProps) {
-  const { register: signup, authError } = useAuth();
-  const { achievement } = useCelebration();
-  const [showPassword, setShowPassword] = useState(false);
-
-  const { register, submit, formState, isSubmitting, watch } = useApiForm<
-    RegisterValues,
-    boolean
-  >({
-    schema: registerSchema,
-    defaultValues: { email: "", password: "", name: "" },
-    onSubmit: async (values) => {
-      const fallbackName = values.email.split("@")[0] ?? "";
-      const ok = await signup(
-        values.email,
-        values.password,
-        values.name?.trim() || fallbackName,
-      );
-      if (!ok) {
-        // Сценарій «вже зареєстровано» обробляє AuthPage (auto-switch
-        // на login). Інші помилки відображає `authError` нижче.
-        if (authError && /вже зареєстровано/i.test(authError)) {
-          onAlreadyRegistered();
-        }
-        throw new Error("");
-      }
-      return ok;
-    },
-    onSuccess: (_ok, values) => {
-      const name = values.name?.trim() || values.email.split("@")[0] || "";
-      achievement(
-        `Готово, ${name}!`,
-        "Твої дані тепер з тобою на всіх пристроях.",
-        [
-          { icon: "🔐", label: "Захищений акаунт" },
-          { icon: "🔄", label: "Синхронізація" },
-        ],
-      );
-    },
-  });
-
-  const passwordValue = watch("password") ?? "";
-
-  return (
-    <form onSubmit={submit} noValidate className="space-y-4">
-      <div>
-        <label
-          htmlFor="auth-name"
-          className="block text-style-caption text-muted mb-1.5"
-        >
-          Ім{"'"}я
-        </label>
-        <Input
-          id="auth-name"
-          type="text"
-          placeholder={"Твоє ім'я"}
-          autoComplete="name"
-          error={!!formState.errors.name}
-          aria-invalid={!!formState.errors.name}
-          disabled={isSubmitting}
-          {...register("name")}
-        />
-        <FieldError message={formState.errors.name?.message} />
-      </div>
-
-      <div>
-        <label
-          htmlFor="auth-email"
-          className="block text-style-caption text-muted mb-1.5"
-        >
-          Email
-        </label>
-        <Input
-          id="auth-email"
-          type="email"
-          placeholder="email@example.com"
-          autoComplete="email"
-          error={!!formState.errors.email}
-          aria-invalid={!!formState.errors.email}
-          disabled={isSubmitting}
-          {...register("email")}
-        />
-        <FieldError message={formState.errors.email?.message} />
-      </div>
-
-      <div>
-        <label
-          htmlFor="auth-password"
-          className="block text-style-caption text-muted mb-1.5"
-        >
-          Пароль
-        </label>
-        <div className="relative">
-          <Input
-            id="auth-password"
-            type={showPassword ? "text" : "password"}
-            placeholder="Мінімум 10 символів"
-            autoComplete="new-password"
-            className="pr-10"
-            error={!!formState.errors.password}
-            aria-invalid={!!formState.errors.password}
-            disabled={isSubmitting}
-            {...register("password")}
-          />
-          <PasswordVisibilityToggle
-            visible={showPassword}
-            onToggle={() => setShowPassword((v) => !v)}
-          />
-        </div>
-        <FieldError message={formState.errors.password?.message} />
-        <PasswordStrengthBar password={passwordValue} />
-      </div>
-
-      {authError && (
-        <div
-          role="alert"
-          className="text-xs text-error bg-error/10 border border-error/20 rounded-xl px-4 py-2.5"
-        >
-          {authError}
-        </div>
-      )}
-
-      <Button
-        type="submit"
-        variant="primary"
-        size="lg"
-        loading={isSubmitting}
-        className="w-full"
-      >
-        {isSubmitting ? messages.loadingActions.registering : "Зареєструватися"}
-      </Button>
-    </form>
-  );
-}
+import { ForgotPasswordPanel } from "./ForgotPasswordPanel";
+import { GoogleSignInButton } from "./GoogleSignInButton";
+import { LoginForm } from "./LoginForm";
+import { RegisterForm } from "./RegisterForm";
+import { useForgotPassword } from "./useForgotPassword";
 
 interface AuthPageProps {
   onContinueWithoutAccount?: () => void;
 }
 
 export function AuthPage({ onContinueWithoutAccount }: AuthPageProps) {
-  const { loginWithGoogle, requestPasswordReset, authError, setAuthError } =
-    useAuth();
+  const { loginWithGoogle, authError, setAuthError } = useAuth();
   const { CelebrationComponent } = useCelebration();
   const [mode, setMode] = useState<"login" | "register">("login");
   const [googleLoading, setGoogleLoading] = useState(false);
-  const [showForgot, setShowForgot] = useState(false);
-  // "idle" → панель рендерить reset-форму; "sending" — кнопка
-  // disabled під час запиту; "sent" — заміняє форму нейтральним
-  // confirmation (без enumeration hint-у), щоб користувач знав
-  // перевірити інбокс.
-  const [forgotState, setForgotState] = useState<"idle" | "sending" | "sent">(
-    "idle",
-  );
-  const [forgotEmail, setForgotEmail] = useState("");
+  const forgot = useForgotPassword();
 
   const switchMode = () => {
     setMode((m) => (m === "login" ? "register" : "login"));
     setAuthError(null);
-    setShowForgot(false);
-    setForgotState("idle");
-    setForgotEmail("");
-  };
-
-  const handleForgotSubmit = async (e: SyntheticEvent) => {
-    e.preventDefault();
-    const target = (forgotEmail || "").trim();
-    if (!target) {
-      setAuthError("Введи email, на який відправити лист.");
-      return;
-    }
-    setForgotState("sending");
-    const ok = await requestPasswordReset(target);
-    setForgotState(ok ? "sent" : "idle");
+    forgot.closePanel();
+    forgot.setForgotEmail("");
   };
 
   const handleGoogleSignIn = async () => {
@@ -455,41 +35,6 @@ export function AuthPage({ onContinueWithoutAccount }: AuthPageProps) {
     // виконається. Скидаємо локальний спіннер тільки на випадок, якщо
     // OAuth не запустився (помилка вже в `authError`).
     setGoogleLoading(false);
-  };
-
-  const openForgotPanel = (currentEmail: string) => {
-    setAuthError(null);
-    setForgotState("idle");
-    setForgotEmail((cur) => cur || currentEmail || "");
-    setShowForgot((v) => !v);
-  };
-
-  // Авто-згортання forgot-панелі після успіху (UX roast 2026-Q2 A14):
-  // confirmation-параграф висить безкінечно без цього — юзер не
-  // розуміє, що робити далі. Після 6 сек бездіяльності закриваємо
-  // панель і повертаємо логін-форму як default state. Кнопка «Назад до
-  // входу» дає ручний вихід раніше (`closeForgotPanel`).
-  const autoCloseTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  useEffect(() => {
-    if (showForgot && forgotState === "sent") {
-      autoCloseTimerRef.current = setTimeout(() => {
-        setShowForgot(false);
-        setForgotState("idle");
-        setAuthError(null);
-      }, 6000);
-    }
-    return () => {
-      if (autoCloseTimerRef.current) {
-        clearTimeout(autoCloseTimerRef.current);
-        autoCloseTimerRef.current = null;
-      }
-    };
-  }, [showForgot, forgotState, setAuthError]);
-
-  const closeForgotPanel = () => {
-    setShowForgot(false);
-    setForgotState("idle");
-    setAuthError(null);
   };
 
   const onAlreadyRegistered = () => {
@@ -539,8 +84,8 @@ export function AuthPage({ onContinueWithoutAccount }: AuthPageProps) {
             {mode === "login" ? (
               <LoginForm
                 key="login"
-                onForgotPassword={openForgotPanel}
-                showForgot={showForgot}
+                onForgotPassword={forgot.openPanel}
+                showForgot={forgot.showForgot}
               />
             ) : (
               <RegisterForm
@@ -549,72 +94,8 @@ export function AuthPage({ onContinueWithoutAccount }: AuthPageProps) {
               />
             )}
 
-            {showForgot && mode === "login" && (
-              <div
-                role="group"
-                aria-label="Скидання пароля"
-                className="text-xs text-text bg-brand-500/10 border border-brand-500/30 rounded-xl px-4 py-3 leading-relaxed space-y-2"
-              >
-                {forgotState === "sent" ? (
-                  <div className="space-y-3">
-                    <p>
-                      Якщо такий email зареєстровано — ми відправили лист із
-                      посиланням для скидання пароля. Перевір вхідні та папку
-                      «Спам». Локальні дані на пристрої залишаються без змін.
-                    </p>
-                    <Button
-                      type="button"
-                      variant="secondary"
-                      size="md"
-                      onClick={closeForgotPanel}
-                      className="w-full"
-                    >
-                      Назад до входу
-                    </Button>
-                  </div>
-                ) : (
-                  <>
-                    <p>
-                      Введи email акаунту — пришлемо посилання для скидання
-                      пароля. Локальні дані на пристрої залишаються без змін.
-                    </p>
-                    <label
-                      htmlFor="auth-forgot-email"
-                      className="block text-style-caption text-muted"
-                    >
-                      Email для скидання
-                    </label>
-                    <Input
-                      id="auth-forgot-email"
-                      type="email"
-                      value={forgotEmail}
-                      onChange={(e) => setForgotEmail(e.target.value)}
-                      placeholder="email@example.com"
-                      autoComplete="email"
-                    />
-                    {authError && (
-                      <p
-                        role="alert"
-                        className="text-error text-meta font-medium"
-                      >
-                        {authError}
-                      </p>
-                    )}
-                    <Button
-                      type="button"
-                      variant="secondary"
-                      size="md"
-                      loading={forgotState === "sending"}
-                      onClick={handleForgotSubmit}
-                      className="w-full"
-                    >
-                      {forgotState === "sending"
-                        ? "Надсилаю…"
-                        : "Надіслати лист"}
-                    </Button>
-                  </>
-                )}
-              </div>
+            {forgot.showForgot && mode === "login" && (
+              <ForgotPasswordPanel state={forgot} authError={authError} />
             )}
 
             {/* eslint-disable-next-line sergeant-design/no-eyebrow-drift --
@@ -627,35 +108,10 @@ export function AuthPage({ onContinueWithoutAccount }: AuthPageProps) {
               <span className="flex-1 h-px bg-line" />
             </div>
 
-            <Button
-              type="button"
-              variant="secondary"
-              size="lg"
-              className="w-full"
+            <GoogleSignInButton
               loading={googleLoading}
-              disabled={googleLoading}
               onClick={handleGoogleSignIn}
-            >
-              <svg width="18" height="18" viewBox="0 0 24 24" aria-hidden>
-                <path
-                  d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-                  fill="#4285F4"
-                />
-                <path
-                  d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-                  fill="#34A853"
-                />
-                <path
-                  d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l3.66-2.84z"
-                  fill="#FBBC05"
-                />
-                <path
-                  d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-                  fill="#EA4335"
-                />
-              </svg>
-              Увійти через Google
-            </Button>
+            />
 
             <div className="text-center pt-1">
               <button
